plot simtel events, refactor
