+++ src/simtools/visualization/plot_simtel_events.py
@@ -1,17 +1,24 @@
 #!/usr/bin/python3
+# isort: skip_file
 """Plots for light emission (flasher/calibration) sim_telarray events."""
 
 import logging
+from pathlib import Path
 
 import astropy.units as u
-import matplotlib.pyplot as plt
-import numpy as np
 from ctapipe.calib import CameraCalibrator
 from ctapipe.io import EventSource
 from ctapipe.visualization import CameraDisplay
+import matplotlib.pyplot as plt
+import numpy as np
 from scipy import signal as _signal
 
+from simtools.corsika.corsika_histograms_visualize import save_figs_to_pdf
+from simtools.data_model.metadata_collector import MetadataCollector
+from simtools.io import io_handler
+
 __all__ = [
+    "generate_and_save_plots",
     "plot_simtel_event_image",
     "plot_simtel_integrated_pedestal_image",
     "plot_simtel_integrated_signal_image",
@@ -28,6 +35,18 @@ AXES_FRACTION = "axes fraction"
 NO_R1_WAVEFORMS_MSG = "No R1 waveforms available in event"
 TIME_NS_LABEL = "time [ns]"
 R1_SAMPLES_LABEL = "R1 samples [d.c.]"
+
+# Choices understood by the dispatcher used below
+PLOT_CHOICES = {
+    "event_image": "event_image",
+    "time_traces": "time_traces",
+    "waveform_matrix": "waveform_matrix",
+    "step_traces": "step_traces",
+    "integrated_signal_image": "integrated_signal_image",
+    "integrated_pedestal_image": "integrated_pedestal_image",
+    "peak_timing": "peak_timing",
+    "all": "all",
+}
 
 
 def _compute_integration_window(
@@ -814,3 +833,189 @@ def _plot_simtel_integrated_image(
     ax.set_axis_off()
     fig.tight_layout()
     return fig
+
+
+def _save_png(fig, out_dir: Path, stem: str, suffix: str, dpi: int):
+    """Save figure as PNG to out_dir.
+
+    Errors during saving are logged as warnings.
+    """
+    png_path = out_dir.joinpath(f"{stem}_{suffix}.png")
+    try:
+        fig.savefig(png_path, dpi=int(dpi), bbox_inches="tight")
+    except Exception as ex:  # pylint:disable=broad-except
+        _logger.warning("Failed to save PNG %s: %s", png_path, ex)
+
+
+def _make_output_paths(
+    ioh: io_handler.IOHandler, base: str | None, input_file: Path
+) -> tuple[Path, Path]:
+    """Return (out_dir, pdf_path) based on base name and input file."""
+    out_dir = ioh.get_output_directory(label=Path(__file__).stem)
+    pdf_path = ioh.get_output_file(f"{base}_{input_file.stem}" if base else input_file.stem)
+    pdf_path = Path(f"{pdf_path}.pdf") if Path(pdf_path).suffix != ".pdf" else Path(pdf_path)
+    return out_dir, pdf_path
+
+
+def _call_peak_timing(
+    filename: Path,
+    *,
+    tel_id: int | None = None,
+    sum_threshold: float = 10.0,
+    peak_width: int = 8,
+    examples: int = 3,
+    timing_bins: int | None = None,
+    event_index: int | None = None,
+):
+    """Call plot_simtel_peak_timing while tolerating older signature.
+
+    Returns a matplotlib Figure or None.
+    """
+    try:
+        fig_stats = plot_simtel_peak_timing(
+            filename,
+            tel_id=tel_id,
+            sum_threshold=sum_threshold,
+            peak_width=peak_width,
+            examples=examples,
+            timing_bins=timing_bins,
+            return_stats=True,
+            event_index=event_index,
+        )
+        return fig_stats[0] if isinstance(fig_stats, tuple) else fig_stats
+    except TypeError:
+        return plot_simtel_peak_timing(
+            filename,
+            tel_id=tel_id,
+            sum_threshold=sum_threshold,
+            peak_width=peak_width,
+            examples=examples,
+            timing_bins=timing_bins,
+            event_index=event_index,
+        )
+
+
+def _collect_figures_for_file(
+    filename: Path,
+    plots: list[str],
+    args: dict,
+    out_dir: Path,
+    base_stem: str,
+    save_pngs: bool,
+    dpi: int,
+):
+    """Generate selected plots for a single sim_telarray file.
+
+    Returns a list of figures. If ``save_pngs`` is True, also writes PNGs to ``out_dir``.
+    """
+    figures: list[object] = []
+
+    def add(fig, tag: str):
+        if fig is not None:
+            figures.append(fig)
+            if save_pngs:
+                _save_png(fig, out_dir, base_stem, tag, dpi)
+        else:
+            _logger.warning("Plot '%s' returned no figure for %s", tag, filename)
+
+    plots_to_run = (
+        [
+            "event_image",
+            "time_traces",
+            "waveform_matrix",
+            "step_traces",
+            "integrated_signal_image",
+            "integrated_pedestal_image",
+            "peak_timing",
+        ]
+        if "all" in plots
+        else list(plots)
+    )
+
+    dispatch: dict[str, tuple[object, dict[str, object]]] = {
+        "event_image": (
+            plot_simtel_event_image,
+            {"distance": None, "event_index": None},
+        ),
+        "time_traces": (
+            plot_simtel_time_traces,
+            {"tel_id": None, "n_pixels": 3, "event_index": None},
+        ),
+        "waveform_matrix": (
+            plot_simtel_waveform_matrix,
+            {"tel_id": None, "vmax": None, "event_index": None},
+        ),
+        "step_traces": (
+            plot_simtel_step_traces,
+            {"tel_id": None, "pixel_step": None, "max_pixels": None, "event_index": None},
+        ),
+        "integrated_signal_image": (
+            plot_simtel_integrated_signal_image,
+            {"tel_id": None, "half_width": 8, "event_index": None},
+        ),
+        "integrated_pedestal_image": (
+            plot_simtel_integrated_pedestal_image,
+            {"tel_id": None, "half_width": 8, "offset": 16, "event_index": None},
+        ),
+        "peak_timing": (
+            _call_peak_timing,
+            {
+                "tel_id": None,
+                "sum_threshold": 10.0,
+                "peak_width": 8,
+                "examples": 3,
+                "timing_bins": None,
+                "event_index": None,
+            },
+        ),
+    }
+
+    for plot_name in plots_to_run:
+        entry = dispatch.get(plot_name)
+        if entry is None:
+            _logger.warning("Unknown plot selection '%s'", plot_name)
+            continue
+        func, defaults = entry
+        kwargs = {k: args.get(k, v) for k, v in defaults.items()}
+        fig = func(filename, **kwargs)  # type: ignore[misc]
+        add(fig, plot_name)
+
+    return figures
+
+
+def generate_and_save_plots(
+    simtel_files: list[Path],
+    plots: list[str],
+    args: dict,
+    ioh: io_handler.IOHandler,
+):
+    """Generate plots for files and save a multi-page PDF per input.
+
+    Also writes metadata JSON next to the PDF.
+    """
+    for simtel in simtel_files:
+        out_dir, pdf_path = _make_output_paths(ioh, args.get("output_file"), simtel)
+        figures = _collect_figures_for_file(
+            filename=simtel,
+            plots=plots,
+            args=args,
+            out_dir=out_dir,
+            base_stem=simtel.stem,
+            save_pngs=bool(args.get("save_pngs", False)),
+            dpi=int(args.get("dpi", 300)),
+        )
+
+        if not figures:
+            _logger.warning("No figures produced for %s", simtel)
+            continue
+
+        try:
+            save_figs_to_pdf(figures, pdf_path)
+            _logger.info("Saved PDF: %s", pdf_path)
+        except Exception as ex:  # pylint:disable=broad-except
+            _logger.error("Failed to save PDF %s: %s", pdf_path, ex)
+
+        try:
+            MetadataCollector.dump(args, pdf_path, add_activity_name=True)
+        except Exception as ex:  # pylint:disable=broad-except
+            _logger.warning("Failed to write metadata for %s: %s", pdf_path, ex)
